+++ test/user/user_api_test.cc
@@ -504,6 +504,7 @@ TEST_F(MujocoTest, AttachSame) {
       <key name="two" qpos="2 0" act="2 2 0 0"/>
       <key name="three" qpos="3 0" act="3 3 0 0"/>
       <key name="attached-two-1" qpos="0 2" act="0 0 2 2"/>
+      <key name="attached-three-1" qpos="0 3" act="0 0 3 3"/>
     </keyframe>
   </mujoco>)";
 
@@ -621,6 +622,7 @@ TEST_F(MujocoTest, AttachDifferent) {
 
     <keyframe>
       <key name="attached-two-1" qpos="0 0 0 1 0 0 0 2" act="2 2"/>
+      <key name="attached-three-1" qpos="0 0 0 1 0 0 0 3" act="3 3"/>
     </keyframe>
   </mujoco>)";
 
@@ -742,6 +744,7 @@ TEST_F(MujocoTest, AttachFrame) {
 
     <keyframe>
       <key name="attached-two-1" qpos="0 0 0 1 0 0 0 2" act="2 2"/>
+      <key name="attached-three-1" qpos="0 0 0 1 0 0 0 3" act="3 3"/>
     </keyframe>
   </mujoco>)";