+++ doc/changelog.rst
@@ -9,6 +9,8 @@ General
 ^^^^^^^
 1. Add :ref:`mjSpec` option for creating a texture from a buffer.
 2. :ref:`shellinertia <body-geom-shellinertia>` is now supported by all geom types.
+3. Add support for :ref:`attaching<meAttachment>` keyframes. Note: this only supports keyframe containing qpos, qvel,
+   and act.
 
 Version 3.2.2 (Aug 8, 2024)
 ---------------------------